Tested all numeric types Fixed small errors for Boolean, DoublePrecision and Real

+++ ABulkCopy.APostgres/Column/ColumnTypes/PostgresDoublePrecision.cs
@@ -15,7 +15,7 @@ public class PostgresDoublePrecision: DefaultColumn
 
     public override object ToInternalType(string value)
     {
-        return decimal.Parse(value, CultureInfo.InvariantCulture);
+        return double.Parse(value, CultureInfo.InvariantCulture);
     }
 
     public override Type GetDotNetType()
